requirements.txt
yfinance==0.2.50
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
pyyaml==6.0.2
